patch fixing issue 5
The gen_persistence does not spawn a new process anymore. This
is important because, IMO, it is not necessary to spawn a new
process for each database command. Ok, this might be interesting
in other scenarios, but, after the server send a message it is
put to sleep (blocking)--- waiting for a gen_persistence response. In this way,
I think that we did not get the main advantages of a new process.

Note: The architecture helped a lot with such a modification. In a
first moment, I thought it would be hard to implement it. Later,