Fixed security breach, and changed the document key size to 64.
The hash function that was being used to generate the keys of the authentication and authorization documents was MD5. The problem is that, since it is now simple to find collisions when using MD5, it could be possible to forge another document with the same key. Given this problem, I have changed the hash function that computes the key of the documents to SHA-256.